+++ script-test/src/test/java/com/aplana/sbrf/taxaccounting/form_template/ndfl/primary_rnu_ndfl/v2016/PrimaryRnuNdflScriptTest.java
@@ -506,20 +506,20 @@ public class PrimaryRnuNdflScriptTest extends DeclarationScriptTestBase {
 
     private NdflPersonIncome createNdflPersonIncomes(int row, String operationId) {
         NdflPersonIncome personIncome = new NdflPersonIncome();
-        personIncome.setRowNum(row);
+        personIncome.setRowNum(BigDecimal.valueOf(row));
         personIncome.setOperationId("11111");
         personIncome.setOktmo("oktmo111");
         personIncome.setKpp("kpp111");
         personIncome.setIncomeAccruedDate(parseDate("01.01.2017"));
         personIncome.setPaymentNumber("aaaaaaaa" + row);
-        personIncome.setTaxSumm(122222);
+        personIncome.setTaxSumm(122222L);
         personIncome.setOperationId(operationId);
         return personIncome;
     }
 
     private NdflPersonDeduction createNdflPersonDeduction(int row) {
         NdflPersonDeduction personDeduction = new NdflPersonDeduction();
-        personDeduction.setRowNum(row);
+        personDeduction.setRowNum(BigDecimal.valueOf(row));
         personDeduction.setOperationId("11111");
         personDeduction.setTypeCode("001");
 
@@ -544,7 +544,7 @@ public class PrimaryRnuNdflScriptTest extends DeclarationScriptTestBase {
 
     private NdflPersonPrepayment createNdflPersonPrepayment(int row) {
         NdflPersonPrepayment personPrepayment = new NdflPersonPrepayment();
-        personPrepayment.setRowNum(row);
+        personPrepayment.setRowNum(BigDecimal.valueOf(row));
         personPrepayment.setOperationId("11111");
         personPrepayment.setSumm(1999999L); //по xsd это поле xs:integer
         personPrepayment.setNotifNum("123-456-000");
